requirements.txt
streamlit
pdf2image
python-docx